Price bracketing patch
Updated the logic that calculates the percent of units that are affordable at the top affordability bracket. This was not behaving as expected and produced higher than expected estimates, this now calculates the units within the top bracket by comparing the max price/rent and where this falls within the range compared to the maximum of the range.